requirements.txt
Django==4.0.3
djangorestframework===3.13.1
django-environ==0.8.1
psycopg2-binary==2.9.3
beautifulsoup4==4.11.1
requests==2.27.1
jsonschema==3.2.0

dj-rest-auth[with_social]==2.2.4
django-filter==21.1
